factor out some shared functions
