+++ components/curriculum-module/server.js
@@ -1,14 +1,11 @@
 import { html, nothing } from "lit";
 
-import { ifDefined } from "lit/directives/if-defined.js";
-import { unsafeHTML } from "lit/directives/unsafe-html.js";
-import { html as hh, unsafeStatic } from "lit/static-html.js";
-
 import { Button } from "../button/server.js";
 import NextIcon from "../curriculum/assets/curriculum-next.svg?lit";
 import PrevIcon from "../curriculum/assets/curriculum-prev.svg?lit";
 import {
   addAttrs,
+  renderCurriculumBody,
   renderSidebar,
   renderToc,
   renderTopicIcon,
@@ -51,7 +48,7 @@ export class CurriculumModule extends ServerComponent {
                 ? html`<p class="module-group">${doc.group}</p>`
                 : nothing}
             </header>
-            ${this.renderCurriculumBody(context, doc)}
+            ${renderCurriculumBody(context, doc)}
             ${this.renderPrevNext(context, doc)}
           </article>
           <div class="toc-container">
@@ -63,19 +60,6 @@ export class CurriculumModule extends ServerComponent {
         </main>
       `,
     );
-  }
-
-  /**
-   * @param {import("@fred").Context<import("@rari").CurriculumPage>} _context
-   * @param {import("@rari").CurriculumDoc} doc
-   * @returns {import("lit").TemplateResult | import("lit").nothing}
-   */
-  renderCurriculumBody(_context, doc) {
-    if (!doc?.body) return nothing;
-
-    return html`
-      ${doc.body.map((section) => this.renderSection(_context, section))}
-    `;
   }
 
   /**
@@ -106,47 +90,5 @@ export class CurriculumModule extends ServerComponent {
           : nothing}
       </section>
     `;
-  }
-
-  /**
-   * @param {import("@fred").Context<import("@rari").CurriculumPage>} _context
-   * @param {import("@rari").Section} section - The section object containing about data.
-   * @returns {import("@lit").TemplateResult | nothing} The Lit HTML template for the about section.
-   */
-  renderSection(_context, section) {
-    const { id, title, isH3 } = section.value;
-    let { content } = section.value;
-
-    // Pre-process section content for proper custom element naming.
-    // After yari has been sunset, change in generic-content
-    // and remove these replacements.
-    content = content?.replaceAll("<scrim-inline", "<mdn-scrim-inline");
-    content = content?.replaceAll("</scrim-inline", "</mdn-scrim-inline");
-
-    switch (section.type) {
-      case "browser_compatibility": {
-        return nothing;
-      }
-      case "specifications": {
-        return nothing;
-      }
-      default: {
-        const level = isH3 ? 3 : 2;
-        if (!id) {
-          return html`<div className="section-content">
-            ${unsafeHTML(content)}
-          </div>`;
-        }
-
-        const heading = hh`<${unsafeStatic("h" + level)} id=${ifDefined(id)}><a class="heading-anchor" href="#${id}">${title}</a></${unsafeStatic("h" + level)}>`;
-
-        return html`
-          <section aria-labelledby=${ifDefined(id ?? undefined)}>
-            ${heading}
-            <div class="section-content">${unsafeHTML(content)}</div>
-          </section>
-        `;
-      }
-    }
   }
 }
