+++ components/curriculum-overview/server.js
@@ -1,17 +1,14 @@
 import { html, nothing } from "lit";
 
-import { ifDefined } from "lit/directives/if-defined.js";
-import { unsafeHTML } from "lit/directives/unsafe-html.js";
-import { html as hh, unsafeStatic } from "lit/static-html.js";
-
 import { Button } from "../button/server.js";
 import NextIcon from "../curriculum/assets/curriculum-next.svg?lit";
 import PrevIcon from "../curriculum/assets/curriculum-prev.svg?lit";
 import {
   addAttrs,
+  renderCurriculumBody,
+  renderModulesList,
   renderSidebar,
   renderToc,
-  renderTopicIcon,
   topic2css,
 } from "../curriculum/utils.js";
 import { PageLayout } from "../page-layout/server.js";
@@ -47,12 +44,12 @@ export class CurriculumOverview extends ServerComponent {
             <header>
               <h1><span>${coloredTitle}</span> ${restTitle}</h1>
             </header>
-            ${this.renderCurriculumBody(context, doc)}
+            ${renderCurriculumBody(context, doc)}
             ${doc?.modules && doc.modules.length > 0
               ? html`
                   <section class="module-contents">
                     <h2>Module list</h2>
-                    ${this.renderModulesList(context, doc.modules)}
+                    ${renderModulesList(context, doc.modules)}
                   </section>
                 `
               : nothing}
@@ -67,97 +64,6 @@ export class CurriculumOverview extends ServerComponent {
         </main>
       `,
     );
-  }
-
-  /**
-   * @param {import("@fred").Context<import("@rari").CurriculumPage>} context
-   * @param {import("@rari").CurriculumDoc} doc
-   * @returns {import("lit").TemplateResult | import("lit").nothing}
-   */
-  renderCurriculumBody(context, doc) {
-    if (!doc?.body) return nothing;
-
-    return html`
-      ${doc.body.map((section) => this.renderSection(context, section))}
-    `;
-  }
-
-  /**
-   * @param {import("@fred").Context<import("@rari").CurriculumPage>} _context
-   * @param {import("@rari").Section} section - The section object containing about data.
-   * @returns {import("@lit").TemplateResult | nothing} The Lit HTML template for the about section.
-   */
-  renderSection(_context, section) {
-    const { id, title, isH3 } = section.value;
-    let { content } = section.value;
-
-    // Pre-process section content for proper custom element naming.
-    // After yari has been sunset, change in generic-content
-    // and remove these replacements.
-    content = content?.replaceAll("<scrim-inline", "<mdn-scrim-inline");
-    content = content?.replaceAll("</scrim-inline", "</mdn-scrim-inline");
-
-    switch (section.type) {
-      case "browser_compatibility": {
-        return nothing;
-      }
-      case "specifications": {
-        return nothing;
-      }
-      default: {
-        const level = isH3 ? 3 : 2;
-        if (!id) {
-          return html`<div className="section-content">
-            ${unsafeHTML(content)}
-          </div>`;
-        }
-
-        const heading = hh`<${unsafeStatic("h" + level)} id=${ifDefined(id)}><a class="heading-anchor" href="#${id}">${title}</a></${unsafeStatic("h" + level)}>`;
-
-        return html`
-          <section aria-labelledby=${ifDefined(id ?? undefined)}>
-            ${heading}
-            <div class="section-content">${unsafeHTML(content)}</div>
-          </section>
-        `;
-      }
-    }
-  }
-
-  /**
-   * @param {import("@fred").Context<import("@rari").CurriculumPage>} context
-   * @param {import("@rari").CurriculumIndexEntry[]} modules
-   * @returns {import("lit").TemplateResult | import("lit").nothing}
-   */
-  renderModulesList(context, modules) {
-    if (!modules || modules.length === 0) {
-      return nothing;
-    }
-    return html`
-      <ol class="modules-list">
-        ${modules.map(
-          (module, j) => html`
-            <li
-              key="ml-${j}"
-              class="module-list-item topic-${topic2css(module.topic)}"
-            >
-              <a href=${module.url}>
-                <header>
-                  ${module.topic
-                    ? renderTopicIcon(context, module.topic)
-                    : nothing}
-                  <span>${module.title}</span>
-                </header>
-                <section>
-                  <p>${module.summary}</p>
-                  <p>${module.topic}</p>
-                </section>
-              </a>
-            </li>
-          `,
-        )}
-      </ol>
-    `;
   }
 
   /**
